refactor: introduce `RCTPlatformImage` (1/2) (#2783)
This PR is split out from #2766 to help make it smaller.

Previously we had the code `@compatibility_alias UIImage NSImage` which
is basically an objective C runtime way of doing `#define UIImage
NSImage`. We don't quite like this, since we are defining the symbol
`UIImage` in an Appkit context. We have a better pattern we can copy
from RCTPlatformView --> RCTPlatformImage. The `RCTPlatform` prefix
basically means it's NS<blah> on Appkit and UI<blah> on UIKit. This is
as opposed to something like RCTUIView, where on iOS it's a typedef, but
on macOS it's a subclass that adds stuff to NSView to make it more
compatible with UIView.

All in all, this should mostly be a rigorous find/replace to make the
code slightly more strict, but have no behavioral changes

RNTester runs locally.